requirements.txt
langchain==0.3.14
langchain-community==0.3.13
langchain-core==0.3.29
llama_cpp_python==0.3.5
faiss-gpu-cu12==1.9.0.post1
streamlit==1.41.1
pypdf==5.1.0
six==1.17.0
